Passing along paramter type override during serialization
We want to ensure serialization occours as the parameter type hint
defined. An example of needing this behavior is a response that gets
wrapped in a data bag. Gson's `@VirtualProperty` could be used on a
sub-class to automatically unwrap the response. Sending that object back
is problematic if the type is guessed based on the provided object
because the request will be wrapped in the same data bag.

Signed-off-by: Nate Brunette <nate.brunette@wheniwork.com>